pyproject.toml
[tool.poetry]
name = "spotion"
version = "0.1.0"
description = ""
authors = ["ClaasJG"]

[tool.poetry.dependencies]
attrs = "^20.2.0"
python = "^3.10"
click = "^8.1.3"
click_params = "^0.3.0"
spotipy = "^2.19.0"
rich = "^12.5.1"
rich-click = "^1.5.1"
yt-dlp = "^2022.6.29"
eyed3 = "^0.9.6"
Pillow = "^9.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
